requirements.txt
disnake
python-dotenv
